♻️Refactor: Move KafkaConsumer to interface layer and Externalize test topic names
- KafkaProducer 와 KafkaConsumer 이름 변경
- KafkaConsumer 를 interface layer 로 위치 이동
- Kafka 테스트 토픽 이름 하드코딩 제거하고 application.yml 에 key, value 셋팅해서 적용
