+++ src/main/java/com/jane/ecommerce/interfaces/listener/KafkaConsumer.java
@@ -1,4 +1,4 @@
-package com.jane.ecommerce.infrastructure.messaging;
+package com.jane.ecommerce.interfaces.listener;
 
 import lombok.extern.slf4j.Slf4j;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Service;
 
 @Slf4j
 @Service
-public class KafkaConsumerService {
+public class KafkaConsumer {
     @KafkaListener(topics = "my-topic", groupId = "${spring.application.name}")
     public void consumeMessage(ConsumerRecord<String, Object> record) {
         log.info("Received message: {} with key: {}", record.value(), record.key());
